Deterministic experiments, improved file handling
Before this commit, there was *four* different source of randomness:
- the random module
- the np.random module
- the libc.rand() function
- the order of results with multiprocessing.imap_unordered

And there was one source of non-randomness: every worker process spawned
during `imap_unordered` was sharing the same random seed.

This commit ruthlessly remove uses of `random` and `libc.rand()`; all
random calls are made with `np.random` (the most flexible of the three).
`imap_unordered` is replaced by `imap`: there was no reason to use the
former, as we are not consumming the results as soon as they are produced.

The code now generates independent random seed for each job given to the
process pool, and the pool size is left to `mutlitprocessing.cpu_count()`.

This commit also introduce a better handling of filepaths, with the
possibility to run the scripts from anywhere (filepaths are relative to
the script path by default). This is useful for `test_reproduce.py`. The
directory containing the results are not constrained to be called `data`
anymore to be created automatically.

On a minor note, several `print` calls have been moved to the  `.format`
syntax.